nedoredesign:(

+++ app/Http/Controllers/OrderController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Order;
+use App\Models\ProductsVariants;
 use Illuminate\Http\Request;
 use App\Http\Requests\OrderRequest;
 
@@ -15,7 +16,7 @@ class OrderController extends Controller
      */
     public function index()
     {
-        $orders = Order::with('status', 'user', 'productsvariants', 'paymentmethod')->get();
+        $orders = Order::with('status', 'user', 'paymentmethod')->get();
         return response()->json($orders, 200, array('Content-Type' => 'application/json;charset=utf8'), JSON_UNESCAPED_UNICODE);
     }
 
@@ -39,7 +40,6 @@ class OrderController extends Controller
     {
         $validated = $request->validated();
         Order::create([
-            'productsvariants_id' => $validated['productsvariants_id'],
             'user_id' => $validated['user_id'],
             'status_id' => $validated['status_id'],
             'paymentmethod_id' => $validated['paymentmethod_id'],
@@ -92,6 +92,11 @@ class OrderController extends Controller
      */
     public function destroy($id)
     {
+        if(ProductsVariants::where('order_id', $id)->first()) {
+            return response()->json([
+                'message' => 'Заказ номер '.$id.' не получилось удалить. В заказе присутствуют купленные товары.',
+            ]);
+        }
         Order::destroy($id);
         return response()->json([
             'message' => 'Заказ номер '.$id.' удален успешно (я пришел с сервера)',
